requirements.txt
numpy==1.18.1
pandas==1.0.1
Pillow==6.0.0
torchvision==0.5.0
torch==1.4.0

